OMAP4: otg: phy: remove uggly mdelay(300)
The original issue with powering on the PHY (and using
300 ms delay after this) is not related to internal
processes in the PHY but is in the incorrect charger
detection future usage.

Now when it is fixed, we can safely remove this uggly
mdelay that sometimes was called in atomic context and
leads to unacceptable system freezing

Tested with different hubs, devices, and booting scenarios.

Change-Id: I54b770d27ab140dd1de4103397881dd3fb807221
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>