gcx: fix mmu context switch issues.
On an mmu context switch, an mmu flush is required.

When the mmu is enabled after a power off, the mmu context
at the head of the cmd buffer queue is made current but the
cached current context was not updated so the next process
to try a commit would throw an l3 bus error.

Change-Id: Ifc5e315e2c237644399db39c60bf8d6dfcd6e491
Signed-off-by: Craig Stout <craig.stout@ti.com>