Tablet2: tc358765: Brightness should not affect gpio 102
Gpio 102 is shared between two functionalities: resetting tc358765 and
backlight enabling. This leads to unwanted tc358765 reset if backlight
is set to 0 from user space. To avoid this situation handling gpio 102
is removed from backlight control. Theoretically backlight should be
sufficiently controlled with PWM and a separate enable pin handling is
not mandatory

Change-Id: Ic7e3a4799cbc966238c7086dd2500d2e2ed0b851
Signed-off-by: Andrii Anisov <andrii.anisov@ti.com>